+++ react/src/Login.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { toast, ToastContainer, Flip } from 'react-toastify';
 import './assets/bootstrap/bootstrap.min.css';